move planner to separate file
